+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.11.7
 - fix: Use rt import path when using beetle bitswap feature.
+- feat: Add in-memory transport.
+- feat: Add options to disable dns transport.
 
 # 0.11.6
 - feat: Add RepoInsertPin::provider and RepoInsertPin::providers. [PR 180](https://github.com/dariusc93/rust-ipfs/pull/180)